Tests/LibWeb/Text/expected/credential-protection-legitimate-login.txt

=== Credential Protection: Legitimate Login Flow Test ===
undefined
Test 1: Form element exists
  form exists: true
  form.action: file:///login
  form.method: POST
undefined
Test 2: Password field exists
  password field exists: true
  password field type: password
  password field has value: true
undefined
Test 3: Form submission is same-origin
  current origin: file://
  form action origin: file://
  is same origin: true
undefined
Test 4: Form has required elements
  username field exists: true
  username field name: username
  password field name: password
  form has 3 elements (including submit): true
undefined
Test 5: Protocol information
  current protocol: file:
  form uses relative path: true
undefined
=== Expected Behavior ===
For a legitimate same-origin form submission:
  - FormMonitor should detect the password field
  - FormMonitor should verify same-origin submission
  - NO security alert should be triggered
  - Form submission should proceed normally
undefined
=== Test Complete ===
This is a legitimate login scenario and should NOT trigger alerts.
